Support Cursor CLI, revise skills UI and improvements markdown loading time
