+++ desktop-shared/src/main/resources/i18n/messages.properties
@@ -1236,7 +1236,7 @@ plans.interactive.skip=Skip (leave blank)
 #Skills
 skills.nav.title=Skills
 skills.view.title=Skills
-skills.view.description=Run AI agents directly on your machine — reading and writing files, executing shell commands, and iterating until the job is done. Powered by {0}. No copy-pasting required.
+skills.view.description=Run AI agents directly on your machine to read and write files, execute shell commands, and get work done. No copy-pasting required.
 skills.view.empty=No skills yet
 skills.view.empty.hint=Create skills in Settings → Skills to get started
 skills.view.run=Run Skill
@@ -1253,14 +1253,15 @@ skills.view.workdir.label=Working directory
 skills.view.workdir.change=Change
 skills.view.workdir.dialog.title=Select Working Directory
 skills.view.workdir.open=Open folder
+skills.view.category.uncategorized=Other
 skills.view.search.placeholder=Search skills ...
 skills.view.context.placeholder=What would you like to do? (optional)
+skills.view.command.mode.hint=Command mode: input starting with / runs as an agent command.
 skills.view.system.prompt.preview=System Prompt
 skills.view.response.title=Response
 skills.view.followup.label=Follow-up
 skills.view.followup.placeholder=Ask a follow-up question or request a change...
 skills.view.followup.send=Send
-skills.view.history.title=History
 skills.view.history.empty=No runs yet
 skills.view.copy=Copy
 skills.view.system.prompt.file=file
