+++ shared/src/main/kotlin/io/askimo/core/skills/agent/ExternalAgentLoader.kt
@@ -23,6 +23,7 @@ object ExternalAgentLoader {
         ClaudeAgent(),
         GeminiAgent(),
         CodexAgent(),
+        CursorAgent(),
     )
 
     /**
